Refactor: reorganize functions
