+++ crates/tool/cooldown_check/src/offender.rs
@@ -1,0 +1,9 @@
+use cargo_metadata::PackageId;
+
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
+pub struct OffenderCrate {
+    pub package_id: PackageId,
+    pub name: String,
+    pub current_version: String,
+    pub minimum_minutes: u64,
+}
